pyproject.toml
[tool.poetry]
name = "dotmac_mkt"
version = "0.1.0"
description = "Marketing & Digital Asset Management for DotMac"
authors = ["DotMac <dev@dotmac.ng>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "0.111.0"
uvicorn = {version = "0.30.1", extras = ["standard"]}
sqlalchemy = "2.0.31"
psycopg = {version = "3.1.19", extras = ["binary"]}
alembic = "1.13.2"
pydantic = "2.7.4"
python-dotenv = "1.0.1"
jinja2 = "3.1.4"
redis = "5.0.4"
celery = {version = "5.4.0", extras = ["redis"]}
prometheus-client = "0.20.0"
httpx = "0.27.0"
google-api-python-client = ">=2.130.0"
google-auth-oauthlib = ">=1.2.0"
itsdangerous = ">=2.2.0"
python-multipart = ">=0.0.9"
opentelemetry-api = "1.26.0"
opentelemetry-sdk = "1.26.0"
opentelemetry-exporter-otlp = "1.26.0"
opentelemetry-instrumentation-fastapi = "0.47b0"
opentelemetry-instrumentation-sqlalchemy = "0.47b0"
opentelemetry-instrumentation-celery = "0.47b0"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
passlib = {version = "1.7.4", extras = ["bcrypt"]}
pyotp = "2.9.0"
cryptography = "42.0.8"

[tool.poetry.group.dev.dependencies]
pytest = "8.2.2"
pytest-cov = "5.0.0"
pytest-asyncio = "^1.3.0"
ruff = ">=0.15.0"
mypy = ">=1.11"
pre-commit = ">=3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ─── Ruff ────────────────────────────────────────────────
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
# E/W: pycodestyle, F: pyflakes, I: isort, B: bugbear, UP: pyupgrade, S: bandit, SIM: simplify
select = ["E", "W", "F", "I", "B", "UP", "S", "T20", "SIM"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E712",   # comparison to True (SQLAlchemy pattern: .is_(True))
    "E711",   # comparison to None (SQLAlchemy pattern: .is_(None))
    "E741",   # ambiguous variable name
    "B008",   # function call in default arg (FastAPI Depends())
    "B904",   # raise without from (too noisy)
    "S101",   # use of assert (ok in tests)
    "S104",   # binding to all interfaces
    "S110",   # try-except-pass
    "UP007",  # Use X | Y for Union (needs from __future__)
    "UP042",  # replace with PEP 604 (needs from __future__)
    "SIM102", # collapsible if (readability)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106", "S105"]  # asserts and hardcoded passwords OK in tests
"scripts/**" = ["T20"]                  # print() OK in scripts
"alembic/**" = ["E501", "F401"]         # auto-generated migration code

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ─── MyPy ────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = false  # gradual adoption

# ─── Coverage ────────────────────────────────────────────
[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/alembic/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ ==",
]
show_missing = true
